+++ water/numbers/format_int.hpp
@@ -1,4 +1,4 @@
-// Copyright 2017 Johan Paulsson
+// Copyright 2017-2018 Johan Paulsson
 // This file is part of the Water C++ Library. It is licensed under the MIT License.
 // See the license.txt file in this distribution or https://watercpp.com/license.txt
 //\_/\_/\_/\_/\_/\_/\_/\_/\_/\_/\_/\_/\_/\_/\_/\_/\_/\_/\_/\_/\_/\_/\_/\_/\_/\_/\_/\_
@@ -73,7 +73,7 @@ class format_int {
 				max_digits = mydigits ? mydigits : static_cast<unsigned>(-1);
 			if(d < 0) {
 				divide *= static_cast<int_>(-1); // * -1 not -divide for visual c++ unsigned warnings
-				d /= static_cast<int_>(mybase) * -1;
+				d /= static_cast<int_>(mybase) * static_cast<int_>(-1);
 				if(d && max_digits != 1) {
 					divide *= static_cast<int_>(mybase);
 					++digits;
